bricks/ev3: use .elf instead of uImage
Replace creating a uImage with using the .elf file directly for the EV3
firmware. This simplifies the build process and avoids the need for
u-boot-tools. U-Boot already knows how to load the .elf file, so we
can use it directly by using `bootelf` instead of `bootm` (this change
was made in the pybricks/v2.0.0 release of u-boot).

In order to keep the size of the .elf file small, we do not enable
`-ffunction-sections` and `-fdata-sections` for the EV3 firmware.
This avoids the large tables of section names that are generated by
these flags, which would otherwise increase the size of the .elf file
by nearly 200 kB. The .elf file is still a bit larger than the uImage
because of a bit more overhead, but only by less than 4kB.