pbio/platform/ev3/start: disable and flush caches
Add a hack to disable and flush the I and D caches at in the startup
code. Previously, when we were using the U-Boot Linux loader to load
our firmware, U-Boot was doing this for us. But since we switched to
a different loader in commit c24798f4b ("bricks/_common/arm_none_eabi.mk:
ev3 uImage is not Linux"), it broke the display among other things.