pbio/sys: Move block device read to driver level.
This moves the platform-specific checksum and saving logic to the driver level.

This is a better fit generally, and it is needed for EV3 where the flash shares
the SPI bus with the ADC driver. When both reside at the pbdrv level, they can
run sequentially without holding up other drivers.

This also renames the public functions to more accurately represent what they
do. This can only be used to save the disk in one go, so we can't call them
generic storage functions.

Fixes https://github.com/pybricks/support/issues/2264