lib/tiam1808: split SPIPinControl() into two functions
Split SPIPinControl() into SPIPinControlGet() and SPIPinControlSet()
functions. Having to pass a pointer to flags to set a value was awkward
and the read/write flag parameter made it difficult to know if we were
reading or writing. And any function that does two completely different
things based on a boolean flag is a code smell.