bricks/media/pb_type_image: Use umm_malloc for image data.
Add support for a new memory allocator (umm_malloc) that is better
for allocating large blocks of memory. Since MicroPython's garbage
collector has to scan every single byte of memory it manages, to check
for pointers, it is not efficient for large blocks of memory that
are not expected to contain pointers. This is especially true for
image data.

So to avoid slow garbage collection/allocation times and filling up
the MicroPython heap, we can use a separate allocator with a separate
heap for image data.