pbio/sys/hmi: Refactor program start logic.
Currently, the UI is split between sys/hmi and sys/bluetooth. This is the first step towards bringing the UI to one place and eliminating the pbsys/bluetooth layer. This will in turn support the clearer separation of the sys/host and ble/usb layers, which will be needed for proper USB support.

This intermediate result works but isn't identical yet to previous behavior. It currently toggles the advertising state but does not fully turn it off or save it persistently.