pbio/sys/host: move stdin buffer to host from bluetooth
Move the stdin buffer from pbsys/bluetooth to pbsys/host so that it can
be used by other host interfaces, such as USB stdio. All connections
will share the same stdin buffer.

All function names are renames from "rx" to "stdin" to reflect the
current usage. The "rx" naming was a holdover from when this was
originally only used for Bluetooth RX via the Nordic UART Service.

A new PBSYS_CONFIG_HOST_STDIN_BUF_SIZE option is introduced to set a
platform-specific buffer size. Bluetooth-only hubs continue to use the
same size that was used in pbsys/bluetooth. Hubs with USB are set to
64 bytes to match the USB endpoint size.