pbio/sys/host: fix pbsys_host_stdout_write()
Fix issue with pbsys_host_stdout_write() not being cancellable. The
static state variables in the function required that the function be
called repeatedly until it returned something other than PBIO_ERROR_AGAIN.
However, in MicroPython, an exception may be raised in between calls,
which would leave the state variables in an inconsistent state causing
incorrect behavior on the next call.

We can save code size on platforms that only have one transport by only
calling the relevant function instead of trying to duplicate the stream
to a non-existent transport.

Then, to handle platforms with multiple transports, rework the logic to
be able to send data to multiple transports without having to maintain
state between calls. This is done by ensuring that we only send the
same amount of data to each transport.