pbdrv/usb_stm32: Implement event subscription
Implement the Pybricks Profile event subscription flag.

This saves us from trying to send messages to the host when there is no
app running to receive them. Previously, trying to write to stdout
would block forever because of this.

The 5 millisecond timeout is arbitrary. Wireshark shows that the OUT
endpoint is being serviced in less than 1 ms, so this should be plenty
of time even if some hosts are an order of magnitude slower. We don't
want to make it too big because it will delay user programs for this
long and we don't want to mess up someone's control loop too badly.