pbio/sys/storage_data: fix strict aliasing issues
Fix strict aliasing issues between block drivers and pbsys storage.

Casting a uint8_t* to a struct pointer violates strict aliasing rules,
which can lead to hard-to-find bugs. We need to ensure that the memory
being pointed to has the same alignment, etc. as the struct it is being
cast to. The usual way to do this is to use a union of the struct and
the other type we want to use (in this case, uint8_t array).

To do this, we need to move the storage data struct definition to a
header file that can be included by both pbdrv/block_device and
pbsys/storage.

Additionally, in block_device_test, the size of the ramdisk block was
wrong and was causing the NXT to crash sometime after starting the REPL.
It also incorrectly had the .noinit section attribute, but depends on
being initialized to zero, so that is removed.

Fixes: https://github.com/pybricks/support/issues/2272
Fixes: ce4253aec ("pbio/sys: Move block device read to driver level.")