pbio/drv/usb/stm32: rework Pybricks USB characteristic handling
The main change here is switching from using BOS descriptors to using
interface/class-specific descriptors for the equivalent of reading BLE
GATT characteristics over USB. BOS are really meant for operating system
capabilities, not application-level protocols.

Instead, we now have interface/class-specific responses to control
requests that allow reading each characteristic individually. Since the
wValue parameter is 16 bits, it is natural to use the BLE 16-bit UUIDs
for this purpose. To avoid potential clashes, we also use bRequest to
sort these by groups. For now, there is one for standard BLE 16-bit
UUIDs and one for Pybricks-defined 16-bit UUIDs.

Now, all remaining BOS descriptor info is known at compile time, so we
no longer need to fill it in at runtime.

There are also some fixes to make the BOS descriptors work correctly
with Chromium-based browsers. Chromium seems to be a bugging and expects
the WebUSB descriptor to be the first device capability and the Microsoft
OS 2.0 descriptor to be the last device capability. Otherwise, it doesn't
seem to parse them correctly.

Also, we had the bcdUSB incorrectly set to 2.0.1 instead of 2.1.0.