pbio/platform/ev3: Enable 5x5 grid UI.
This enables the SPIKE Prime UI on EV3 by mimicking a 5x5 grid display.

This unblocks the path to implementing USB download and run on EV3 without having to decide or spend time on the UI yet.

For now, attempting to run slot 0 will always start the UART REPL so that other developments can still be worked on in the mean time.