pbio/drv/usb_ev3: fix USB not working after firmware update
Fix the USB being in a bad state after a firmware update.

The firmware updater uses the USB to transfer the new firmware and then
jumps to the new firmware without resetting the USB controller. Calling
USBReset() was not enough to reset the USB controller. Even after this
we would never get any interrupts, so requests from a host would time
out. This is fixed by using the PSC to reset the USB controller instead
of USBReset().

Also, before we reset things, we need to tell the host to disconnect
from the firmware update USB device. When we reconfigure things the EV3
will appear as a completely different USB device.

Fixes: https://github.com/pybricks/support/issues/2295