pbio/drv/usb/usb_ev3.c: Implement status keepalive timer
Without this timer, if the hub state doesn't change and the user program
doesn't output anything, the hub won't be able to detect that the
host software has closed (because it never sends anything which could time out).

Keeping this state reasonably up-to-date is important for the hub
auto-power-off functionality.