pbio/drv/ioport: Allow IO port VCC control to be optional
The EV3 technically has an IO port power control, but it doesn't only
control the main peripheral ports but also the ADC and USB.
It also doesn't require quirk handling to be able to power-cycle the ports.

Make this feature compile-time optional rather than having dummy values.
This was previously silently dereferencing a null pointer.