pbio/platform/ev3/platform.c: Initial MMU setup
This creates an identity mapping for only the valid bits of
the address space. It also adds logic to the exception handler
to print out fault addresses.

This currently marks the entire address space as uncacheable.
Drivers will need to be updated to take caches into account
before we can mark RAM as cacheable.