bricks/_common/mphalport: use same file for EV3 and STM32
Move the STM32 mphalport.c to the common directory so that it can be
used for both the EV3 and STM32 ports. Eventually, all platforms should
use the same file. The EV3 clock driver had to have a small change to
make it work with the common mphalport.c.

Also clean up some duplication with regard to thumb assembly code files
while we are touching this.