pbio/port: Reset power to port on disconnect.
We used to turn power off only at the end of active sensor processes as well as at the end of a program. In general, we want to turn power off whenever something is disconnected, which is what we do here.

The goal here is to be able to turn off 9V to a sensor if a user has switched it on and then unplugs the device.