pbio/port_dcm_ev3: Drop outer loop.
This is only an indentation change with continue replaced by returning success.

This thread should wait for a connection and then wait until disconnected. The outer sensor port loop will run this again instead.

This allows us to handle things like port reset in the outer sensor loop. This ensures 9V power is turned off once a device is unplugged, for example.