pybricks.nxtdevices.UltrasonicSensor: Implement in C.
Given the firmware size limitations on NXT, using Python wasn't going to be scalable. With the recent changes to the I2CDevice implementation, we can use instances of it for sensor classes written in C.

This reverts commit a37f46f756cd8453ff8f99193263ac68f9b5a057.
This is reverted instead of dropped so we maintain the logic to import Python classes into the C module in git history for future reference.