pbdrv: adc: ev3: Improve ADC settling time.
Using the max SPI SCLK speed of 20 MHz, the ADC is not able to read up
to the full 5V but rather reading somewhere around 50 mV less.

By slowing down the SCLK rate, we give the ADC more time to settle after
switching the mux and can get more accurate readings as a result.
Testing has shown that 1 MHz slow enough to achieve this.