pbio/sys/light: Move breathe animation.
This animation runs in user space, so it doesn't belong with the status event code. It used to be closely tied to the event messages since this used to be the only way to know that a program was started. Now we can just start the animation when the program starts.

This will also allow the light instance to reset the light to the default behavior.