pbio/sys/light: Don't pass unused status data around.
This was passed to the light status update function, but it does the same
regardless of setting or clearing.

This is not related to the pbsys_status_set function, which takes a bool for
set/clear. So pbsys_status_change_t made things a bit confusing now that it is
no longer used.