pbio: Implement text drawing
Use only one font for the moment, Terminus Font, 8x16 pixels size.

A script is included to generate the font source using freetype-py.

Only handle ASCII characters as discussed.

Antialiasing is not that great with the limited number of gray levels on
the EV3, so use only monochrome font, which works great thanks to
hinting.

Refs: https://github.com/pybricks/support/issues/2154