pbio/drv/bluetooth_stm32_cc2640: Use loops to set params.
When inspecting the map.elf, the init_task was well over 4K. Since much of it is repetition with different parameters, we can vastly reduce the number of different entry and yield points by looping over a list of parameters instead.

This reduces the firmware size by nearly 2K, which is helpful since we are running short on Technic Hub.

Almost everything is still set in the same order. Only the bond manager configuration is reordered slightly so we can use a single loop.