pbio/sys/light_matrix: Fix crash on unfinished boot animation.
If we start a program on boot, the startup animation isn't finished. If we start the run animation then, we get a crash on running pbio_light_animation_stop_all after running the user program.

Plenty of time wasted on this one. There is nothing wrong with the MicroPython exception handler. Thank you for reading.