pbio/drv/bluetooth: Add way to forcefully exit.
Such as mechanism had not yet been implemented after overhauling the Bluetooth drivers.

Before, we used to try to disconnect and stop observing/scanning in the MicroPython deinit, and stop waiting in case of a shutdown request.

No we do it as part of the pbio cleanup, and enforce shutdown if it fails since this is not a recoverable state.