pbio/drv/bluetooth: Maintain advertising state.
We used to maintain an advertising state in pbsys/bluetooth and now in pbsys/hmi, handling things such as _not_ advertising if it was already in the connected state. This works, but becomes even more cumbersome when we add a USB state in the next commits.

It is simpler for the higher level code to be able to just say stop_advertising, and letting the Bluetooth driver treat it as a no-op if it not advertising to begin with.