pbio/drv/bluetooth: Don't wait on power initialization.
Since all Bluetooth commands are buffered and handled by the event loop, we don't need to wait for the driver to fully power on before leaving `pbdrv`.

This allows power on to complete in the background while we initialize pbio and do hmi boot animations in pbsys, saving about half the boot time on SPIKE Prime.