pbio/sys/hmi_pup: Passively monitor connection state.
This way it can run even after the UI completes without needing a separate process. This also simplifies the HMI loop itself, since it won't need to monitor changes.

Since BLE advertising is now a no-op when already in the expected state, we can just keep setting it in the waiting loop.