pbio/sys/storage: Limit program size to slot division.
Technically, only the sum of all slots is limited since they are stored
consecutively. However, we send only one static value to Pybricks Code, so it
doesn't know about the dynamic limit. Having a fixed limit ensures that we
can give the appropriate error message for large programs instead of failing
with a Bluetooth/USB error code when storage is full.

This limits programs on SPIKE Prime to 52,300 bytes, which should cover most
use cases. We can relax this limit once we make upgrades to Pybricks Code.