bricks/_common/micropython: Run first mpy instead of `__main__`
Default to running the first `mpy` file instead of searching for __main__ in
the metadata. Now hosts can specify the actual file name so we can include
it in on-brick UIs like on EV3.

This is distinct from __name__ = "__main__" which will still work, just as it
does in normal Python.

See https://github.com/pybricks/support/issues/2364