bricks: Add simhub for testing embedded ports on CI.
The existing VirtualHub works well enough for MicroPython coverage testing, but it is too far removed from the other Pybricks ports to be used for everyday debugging.

This introduces a new port based on the minimal port for running on Unix systems. Currently supports the REPL and works well with the motor simulation and animation.

It uses a simulated Bluetooth driver to provide stdio, closely mimicing the behavior or of the real hubs. Can also be used to debug importing, running mpy files, and so on.