bricks: PB_LIB_BTSTACK as string variable
The PB_LIB_BTSTACK library has to be compiled
with different sources if it is being used to drive
a classic module vs. a dual or LE module.
Per PR feedback, adjust the PB_LIB_BTSTACK variable
to signal which state we are in, rather than using separate
PB_LIB_BTSTACK and PB_LIB_BTSTACK_CLASSIC variables.