pbio/platform/sim_hub: Don't use Linux signal for clock tick.
This interrupts syscalls like reading data from a pipe, especially when debugging. We can instead just request polling when MicroPython drives the poll hook and the wall clock time has changed.

For the CI clock variant, this was not implented at all yet. While we are at it, ensure it increments the clock every N instructions, so that even tight loops without any waits will advance. For example, this would be needed to run asynchronous scripts that don't _WFI anywhere.