pbio/drv/bluetooth: Don't restart observing too often.
This timer was missed in the overhaul of 023d6f61f9f63532b1c552abf8bd69dbf3267833.

If a user would call observe in a tight loop when there is nothing to receive, this would repeatedly restart scanning.

Since the btstack driver doesn't await most operations, this would lead to a crash eventually.