pbdrv/uart_debug_first_port: fix bugs/va_list func
It's useful to have a vprintf version of
the uart_debug functions. This will be used
in a later change to assist with logging
messages originating in BTStack.

Also, there was a bug where the error from
writing to the UART was not captured. This
fixes that bug and logs the error on
the next attempt to use the port.