bricks/_common/mphalport: Don't WFI on completed prints.
This would cause Bluetooth messages to inefficiently buffer. MicroPython calls mp_hal_stdout_tx_strn at least twice for every print because it makes another call for \r\n. The user may also call it multiple times with short messages.

Running the poll hook even when the buffer is not full meant that it would begin to transmit it rather than concatenate them as we intended (this has been in the driver for years).

It would also hold up the virtual hub, incrementing the clock twice for every print.