pbio/drv/usb: Add common USB driver.
This de-duplicates common USB code such as handling stdout or sending the hub status. This was quite hard to keep track of between the EV3 and STM32 drivers, causing subtle differences and issues which were not so easy to synchronize.

The simulation driver shows the bare minimum driver-specific functions that each driver must implement, which will we do in the next commits.