pybricks/bluetooth: Add bt classic inquiry scans.
Any Pybricks device running btstack can now issue Bluetooth
classic inquiry scans to find and report discoverable devices.

This API is experimental and we can change it later if we
don't like it. But it's nice because it can be used to concretely
demonstrate that the EV3's BTStack layer is doing something!
This also should give some indication of the structure by which
we'll add further classic functionality. If it's preferred to
have it in the same file as the LE code, we can, but this structure
is scarcely more expensive and a little more neatly organized.