pybricks.iodevices.AnalogSensor: Update for embedded.
This remains the same as it was, except we also allow custom devices that aren't detected as NXT analog type.

This can also be used to for devices that would otherwise be detected as an NXT Light or Color Sensor, so as to not run the background process that toggles the light.