pbio/color/util: Add heuristic distance for saturated colors.
The bicone mapping is highly distant dependent, which makes it suitable in
limited cases, and it doesn't work great with the default colors.

If only saturated or grayscale colors are in the mapping, we can generally get
a much better result by looking at hue only for saturated colors and looking at
value only for unsaturated colors, and use the saturation to decide which to
pick. This also means we won't need the workaround of having negative V for None.

The logic here is that if you do specify fine-grained, calibrated colors, then
it will use the original bicone distance mapping.