pybricks.tools: Fix race=False in multitask().
Fix wrong index direction for fetching the race keyword argument in
pb_type_Task_new().

n_args - 1 would be the last positional argument, n_args would be the
race keyword itself and n_args + 1 is the value of the race keyword
argument, which is what we want.

Also add a test to verify that it is working as expected.