pbio/drv/bluetooth_btstack: Don't allow re-entry during power control.
There was a subtle case not handled here. Multiple power control calls were prevented, but it was technically possible for the re-entry to immediately satisfy the PBIO_OS_AWAIT_UNTIL condition, move on with the main process, and trigger some next Bluetooth command.

We shouldn't do that. Instead, ensure that re-entry does not proceed at all, letting only the original call proceed.