pbdrv/uart_debug_first_port: Log on panic.
It's helpful when you crash to see what happens right before you
crash. For example, if you add a log message "about to dereference
sketchy pointer" and then you get a segfault, you'd love to know
whether that message occurred. This change makes it so that at least
on EV3, we will.

This also adds "pb_log" as an alias for pbdrv_uart_debug_printf.
Usually, platforms make it easy to log debug messages, and we should
be no different. C.f. `ESP_LOGI`, Zephyr's `LOG_ERR`, glog's
`LOG(WARNING)` etc.

Tested by adding the following code to pbdrv_init.

```
    pb_log("test\n");

    lwrb_t* rb = (lwrb_t*)0x1281441490;
    pb_log("%p\n", lwrb_get_free(rb));
```

Before this, the "test" would not appear before the panic.
After, it does. If a panic occurs while some messages are pending
in the uart send state, some bytes may be duplicated on the log.