pbio/port_dcm_ev3: Allow use without autodetect.
Instead of attempting to use partial autodetect when full detection is not possible, it is more reliable to just set the intended type when the user initializes a class.

This also simplifies the #if PBDRV_CONFIG_IOPORT_HAS_GPIO_P2 logic to disable all detection code instead of many selective portions.