pbio/drv/display_nxt: Convert to pbio model.
Use a PBIO thread to handle the screen refresh instead of doing it in
ISR.

The driver provides the frame buffer which is used by the user. On
refresh, format is converted to match the one used by the LCD driver.

In the future, the user frame buffer may use the LCD driver format in
order to improve memory usage. This will require pbio/image to handle
this format.

Refs: https://github.com/pybricks/support/issues/2425